+++ src/modules/reportSalaryPartner/components/ModalDetail.tsx
@@ -31,6 +31,7 @@ import { ItemVoucher } from "./Context";
 import WithPermission from "~/components/common/WithPermission";
 import POLICIES from "~/modules/policy/policy.auth";
 import { METHOD_TYPE } from "~/modules/vouchers/constants";
+import { useLocation } from "react-router-dom";
 import { ReportSalaryPartnerProvider } from "../ReportSalaryPartnerProvider";
 type propsType = {
   id?: string;
@@ -205,7 +206,7 @@ export default function ModalDetail(props: propsType): React.JSX.Element {
   const [form] = Form.useForm();
   const [open, setOpen] = useState(false);
   const [openReceipt, setOpenReceipt] = useState(false);
-
+  const { pathname } = useLocation();
   const infoData: any = useMemo(() => {
     return data.find((p: any) => p._id === props?.id);
   }, [data, props?.id]);
@@ -309,43 +310,34 @@ export default function ModalDetail(props: propsType): React.JSX.Element {
           <BoxMoney title={"Tổng đã thu"} total={totalReceipt} />
           <BoxMoney title={"Tổng đã chi"} total={totalPayment} />
         </Flex>
-        
-          <Flex style={{ marginTop: 20 }} justify="end" gap={10} align="center">
-            {/* <WithPermission permission={POLICIES.READ_VOUCHERSALARYPARTNER}> */}
-            <Popover
-              trigger={["click"]}
-              title="Danh sách phiếu"
-              content={
-                <VoucherList dataSource={get(infoData, "vouchers", [])} />
-              }
-            >
-              <Badge count={get(infoData, "vouchers", []).length}>
-                <Button icon={<MenuOutlined />} type="primary" ghost>
-                  Danh sách phiếu
-                </Button>
-              </Badge>
-            </Popover>
-            {/* </WithPermission> */}
-            <ReportSalaryPartnerProvider refCollection={infoData?.typeSaler} methodType={METHOD_TYPE.VOUCHER_SALARY}>
-              <WithPermission permission={POLICIES.WRITE_VOUCHERSALARYPARTNER}>
-            {total === 0 && <Tag color={"success"}>Đã hoàn tất thanh toán</Tag>}
-                {total > 0 && (
-                  <Button type="primary" onClick={onOpenPayment}>
-                    Tạo phiếu chi
-                  </Button>
-                )}
-          
-            {/* <WithPermission permission={POLICIES.WRITE_VOUCHERSALARYPARTNER}> */}
-            {total < 0 && (
-              <Button type="primary" onClick={onOpenReceipt}>
-                Tạo phiếu thu
-              </Button>
-            )}
+      <Flex style={{ marginTop: 20 }} justify="end" gap={10} align='center'>
+        <WithPermission permission={pathname === PATH_APP.reportSalaryPartner.root ? POLICIES.READ_VOUCHERSALARYPARTNER : POLICIES.READ_VOUCHERSALARYEMPLOYEE}>
+        <Popover
+          trigger={["click"]}
+          title="Danh sách phiếu"
+          content={<VoucherList dataSource={get(infoData, "vouchers", [])} />}
+        >
+          <Badge count={get(infoData, "vouchers", []).length}>
+            <Button icon={<MenuOutlined />} type="primary" ghost>
+              Danh sách phiếu
+            </Button>
+          </Badge>
+        </Popover>
+        </WithPermission>
+        <ReportSalaryPartnerProvider refCollection={infoData?.typeSaler} methodType={METHOD_TYPE.VOUCHER_SALARY as any}>
+        {total === 0 && <Tag color={'success'}>Đã hoàn tất thanh toán</Tag>}
+      <WithPermission permission={pathname === PATH_APP.reportSalaryPartner.root ? POLICIES.WRITE_VOUCHERSALARYPARTNER : POLICIES.WRITE_VOUCHERSALARYEMPLOYEE}>
+      {total > 0 && (
+          <Button type="primary" onClick={onOpenPayment}>
+            Tạo phiếu chi
+          </Button>
+        )}
+        {total < 0 && <Button type="primary" onClick={onOpenReceipt}>
+            Tạo phiếu thu
+          </Button>}
             </WithPermission>
-             </ReportSalaryPartnerProvider>
-            {/* </WithPermission> */}
-          </Flex>
-       
+            </ReportSalaryPartnerProvider>
+      </Flex>
       </Flex>
 
       <ModalAnt