+++ src/routes/allPath.ts
@@ -97,7 +97,7 @@ export const PATH_APP = {
     root: path(ROOTS.app, "/vouchers"),
     supplier: path(ROOTS.app, "/vouchers-supplier"),
     pharmacy: path(ROOTS.app, "/vouchers-pharmacy"),
-    salaryPartner: path(ROOTS.app, "/vouchers-salaryPartner"),
+    salary: path(ROOTS.app, "/vouchers-salary"),
     partner: path(ROOTS.app, "/vouchers-partner"),
   },